refactor: move tests to `conversation` module
This is where `get_client_ids()` and `export_secret_key()` are
implemented, so this is the place to be for their unit tests.
